Move get-shebang
