+++ src/common/util.js
@@ -580,17 +580,6 @@ function isFrontMatterNode(node) {
   return node && node.type === "front-matter";
 }
 
-function getShebang(text) {
-  if (!text.startsWith("#!")) {
-    return "";
-  }
-  const index = text.indexOf("\n");
-  if (index === -1) {
-    return text;
-  }
-  return text.slice(0, index);
-}
-
 /**
  * @param {any} object
  * @returns {object is Array<any>}
@@ -664,7 +653,6 @@ module.exports = {
   addDanglingComment,
   addTrailingComment,
   isFrontMatterNode,
-  getShebang,
   isNonEmptyArray,
   createGroupIdMapper,
 };
