+++ src/language-js/parse/babel.js
@@ -3,8 +3,8 @@
 const tryCombinations = require("../../utils/try-combinations.js");
 const {
   getNextNonSpaceNonCommentCharacterIndexWithStartIndex,
-  getShebang,
 } = require("../../common/util.js");
+const { getShebang } = require("../utils/get-shebang.js");
 const createParser = require("./utils/create-parser.js");
 const createBabelParseError = require("./utils/create-babel-parse-error.js");
 const postprocess = require("./postprocess/index.js");
